Project/requirements.txt
openai
python-dotenv
pypdf
tiktoken
langchain
ollama
